Game design   minh tuan

+++ ArchersGame/Lightning/Lightning.hpp
@@ -26,6 +26,6 @@ public:
     void updateFrame(double time);
     void change_volume_lightning(float volume);
 
-    int getLightningRate();
-    void setLightningRate(int rate);
+    float getLightningRate();
+    void setLightningRate(float rate);
 };
